--- src/flex_evals/checks/extended/__init__.py
@@ -1,15 +1,0 @@
-"""
-Extended check implementations for FEP.
-
-Imports all extended checks to trigger their registration.
-"""
-
-from .semantic_similarity import SemanticSimilarityCheck_v1_0_0
-from .llm_judge import LlmJudgeCheck_v1_0_0
-from .custom_function import CustomFunctionCheck_v1_0_0
-
-__all__ = [
-    "CustomFunctionCheck_v1_0_0",
-    "LlmJudgeCheck_v1_0_0",
-    "SemanticSimilarityCheck_v1_0_0",
-]
