+++ tests/test_checks/__init__.py
@@ -1,0 +1,15 @@
+"""Test package for individual check implementations.
+
+This package contains comprehensive tests for each check type,
+organized by functionality rather than implementation details.
+Each test module covers:
+
+- Pydantic validation and field handling
+- Execution logic and __call__ methods  
+- Engine integration tests
+- Edge cases and error handling
+- JSONPath integration tests
+
+All tests preserve the existing functionality and test coverage
+from the previous schema/implementation separation.
+"""
